feat[forrest]: add `map::Map`
This is a verifiable map implementation. It creates a map of keys and
values in a Merkle tree so that inclusion proofs can be generated from
the root hash. The initial implementation is intentionally naive so that
we can agree on the invariants we desire from the structure.

Currently, each map is immutable and is modified by forking the previous
map into a new map. `Rc` is used internally for efficient sharing of
objects between map instances. If we need to share instances across
threads, we can switch to `Arc`.

The Merkle tree itself is sparse and uses a prefixed hashing scheme.
Inclusion proofs are compressed by omitting leading sparse peers.

Signed-off-by: Nathaniel McCallum <nathaniel@profian.com>