Remove the `--queue` option in favor of specific pending publish commands.
This commit removes the `--queue` option from `publish release` and instead
adds back the `publish start` command to start a "pending" publish.

A new `publish init` command was added to initialize a package independently of
a release.

Now the `publish init` and `publish release` commands will enqueue the records
if there is a pending release or otherwise immediately perform the operation if
not.