Implement shared (per-user) package and content storage in the client. (#78)
This PR implements storing package logs and content at a known location
by default (the user's cache directory).

It also implements a simple configuration file format for the client
that can be passed in via a `--config` CLI option or read from known
locations.

A `--registry` option for which registry to communicate with has been
added to the CLI; can be defaulted to a particular registry in the
client configuration.

The `warg-cli` executable has been renamed back to `warg`.

Client storage has been split into two traits: one for package storage
and one for content storage. This allows a shared content storage
directory while having per-registry package storage directories.

By default, the storage layout is now (overridable with a client
configuration):

* `$CACHE_DIR/warg/packages/<registry>/<algo>/<digest>`
* `$CACHE_DIR/warg/content/<algo>/<digest>`
* `$CACHE_DIR/warg/content/temp/...`

The `init` command in the CLI has been removed in favor of a `config`
command for configuring the client.

The demo and README have been updated to reflect these changes.