Use a single reqwest client in API client.
As reqwest's client is an `Arc` around an inner client that maintains a
connection pool internally, it is best (and recommended by reqwest
documentation) to create a client once and reuse it.

This commit moves the reqwest client into the API client as a field rather than
creating the client on demand.