Add configurable data stores to the server.
This (large) commit implements a new `--store` option to the server.

The option defaults to `memory`, which stores logs and checkpoints in memory
which will be lost when the server shuts down.

With the `postgres` feature enabled, it also supports a `postgres` store option
which requires the `DATABASE_URL` environment variable to point at the
PostgresSQL server to use.

The `postgres` store supports restarting the server and recreating the
necessary data structures for the transparency service to work.

See `crates/server/README.md` for information on how to setup a persistent
PostgresQL store.