Add the first server integration tests.
This commit adds the first server integration tests.

There are currently three simple tests:

1) Start a server and verify the creation of the operator log.
2) Start a server and publish a component.
3) Start a server and publish a WIT package.

This also introduces a `Server` to the `warg_server` crate that encapsulates a
lot of the logic that was previously in `main.rs`.

This allows the tests to easily create a server in the same process as the
tests and determine which port the test client should use to communicate with
the server.

Additionally, the `Config` representation for the server was updated to have
options for whatever is currently supported by the CLI.

The checkpoint interval is also now configurable; tests use this to decrease
the time it takes for a publish to checkpoint.

More tests are coming.