Move `Validator` trait to protocol crate.
This commit adds a `Validator` and `Record` trait to the protocol trait.

It enables code to interact with both operator and package log validators in
a generic fashion.

This commit also removes the `unsafe` method for the validators and makes the
transaction mechanism private again.

It is now an error to provide a content source for a digest that isn't needed
to validate a package record.