Add an `openapi.yaml` file to the server implementation.
This commit creates a single OpenAPI specification for the Warg registry
server.

It is intended to be the complete specification of a Warg registry server's
REST API.

The specification proposed here is not what is currently implemented by the
server, but it is inspired heavily by both the existing implementation and
previously proposed APIs.

We can use this as a starting point to finalize an API specification for the
initial version.