Refactoring of REST API, client, and server.
This commit includes a sizable refactoring to the client and server to conform
to the proposed spec in #101.

It includes improvements to the REST API that enable the following features:

* Fetching logs now supports paging for logs with lots of records.
* The server now informs the client of what content it should upload to start
  processing a newly published record, rather than the client preemptively
  sending all the content to the server and giving it URLs.
* The client now downloads a package's content by requesting the record and
  downloading the content via the sources associated with the record, rather
  than having explicit knowledge of where the content is located.

Content uploads are now tied to records that are being published; thus the
content service (which allowed completely arbitrary uploads) has been removed
and the relevant functionality folded into the package service.

Additionally, the REST API was moved under a `v1` path and separated out in
both the client and server for future versions.

CORS was also enabled for the API to allow for it to be used from browsers.

Note: a new DB migration was not created for this work as no one is really
using the database feature just yet; in the future when folks are actually
using the database, a migration would be required.

Closes #104.
Closes #99.