Remove specification of content sources on publish from the API.
This commit attempts to simplify how the content sources are handled.

Right now, externally specified content sources aren't used and ultimately
complicate the design of the initial implementation of the registry.

Namely, content must pass through the server's (forthcoming) policy system; so
that means even if users specify the content sources at publish, the server will
need to download from those URLs (it's not a great idea to have an API where the
user can cause connections from your backend to arbitrary hosts) to ensure the
content passes policy checks.

Further complicating things, processing remote content would likely need to be
done asynchronously in addition to processing the record itself.

That said, even with these changes, servers are still free to store the uploaded
content wherever they wish and associate the URLs it knows with the record.

Registries that mirror other registries can also choose to mirror the content
and provide their own content sources for a record (or otherwise return the
original URLs).

This is a verbose way of saying I'm yeeting specifying external content sources
from the API spec until such time that it is definitely required and we give
proper thought into how it would work.