fix: return bound endpoints after server started
The impetus behind the change is to support adding monitoring endpoints with
an optional additional TCP listener. While refactoring the server it was
discovered that the run method moves the Server self, and so fixed #112 to
also ensure that the API endpoint was running before tests could discover the
listener's locally bound port.

In addition a common cancellation token is added to the server to propagate
shutdown to both the core service as well as the API server endpoint. It can be
later used to support additional services later such as a monitoring
endpoint(s).