Add an explicit 403 response to publishing records.
This commit explicitly adds a 403 response to publishing records.

A server will use this response when the key used to sign the record is not
authorized to publish to the log.

Note that this response will be used in the future when the server does a
policy check to see if the key is permitted to publish to the log's name (i.e.
even for init records as the check is done *before* processing the record).