Add the `unauthorized` response for publishing records.
This commit adds the `unauthorized` response for publishing records to the API
crate.

It is not a currently used response (the server will accept *any* request to
publish provided the record validates), but in the near future this will be the
expected response when the registry is configured to only accept certain
signature keys for log names.