Fix inclusion proof on client with incremental fetch.
The inclusion proof should not be inside the loop for fetching and validating
records.

If a response was received the included `more` set to `true`, the inclusion
proof check would include a log leaf that is not for the checkpoint we're
fetching for, resulting in a inclusion proof failure.

The fix is to simply do one inclusion proof request after all records for the
checkpoint have been fetched and validated.