Add a test for the client for incrementally fetching package records.
This commit adds a test for incrementally fetching package records.

It publishes 2000 release entries, deletes the client storage, and then does an
`upsert` on the client to insert the package log.

As there are over the maximum allowable limit of records in the log, the client
will fetch and validate the log over multiple requests to the server.

This commit includes a minor rename of `max` to `limit` in the server and a
change to the test code to create the storage roots with the test process name
in the path (now that there is more than one).