Implement a WebAssembly content policy in the server.
This commit implements a content policy in the server that incrementally
validates that content uploaded to the server is valid WebAssembly.

Currently the policy is not exposed via the CLI, but the integration tests use
the policy to expect that all content should be WebAssembly.

We may want to expose enabling some policies via the CLI in the future, but for
now this should allow for consumers of the `warg-server` crate to enable
policies for custom server implementations.