Add `--content-base-url` option to `warg-server`.
This commit adds a `--content-base-url` to `warg-server` that allows specifying
what the expected base URL to use when serving content.

By default, the base URL is derived from the listen address for the server.

For servers that aren't bound to the address that will also be serving the
content, the `--content-base-url` option can be used to control the URLs the
server returns for fetching content.