Update crates for publishing. (#207)
This PR is going into the `release-0.1.0` branch (off of the `v0.1.0`
tag), which, when ready, will contain everything for the 0.1.0 crates to
be published.

To start, this updates the various `Cargo.toml` for fields related to
publishing. The next step is creating a publish script for a GitHub
action to auto publish with.

Please 🚲 🏠  on the crate descriptions a bit.