fetch logs endpoint using fetch token String (#203)
This changes the `POST /v1/fetch/logs` endpoint to use a server
implementation defined `fetchToken` for the last known record instead of
the `RecordId`. It still can be the `RecordId`, but it doesn't need to
be. In fact, the reference implementation still uses the `RecordId` for
the `fetchToken`.

The motivation here is to give server implementations greater
flexibility on implementing the fetch logs endpoint.

Also, in this PR, the Client CLI will be more resilient in handling
records that are returned that it has already seen before.