Fix transaction implementation in `LogState::validate`.
This commit removes the incorrect transaction implementation for
`LogState::validate`.

Now `validate` takes ownership of `self` and returns `Result<Self,
ValidationError>`.

This means that callers that expect to keep the log state following an invalid
log entry must clone the state prior to validation.

As the in-memory data store is the only store that persists the log state in
memory, it now clones the state before validation and updates the log state
upon successful validation.

For the postgres data store, the log state was loaded from the database and is
discarded on error, so no clone is necessary.

Fixes #242.