Depsolve client (#199)
This PR adds three commands to the warg CLI, `dependencies`, `lock`, and
`bundle`

- `warg dependencies <namespace:package>`
Recursively fetches based on occurrences of `unlocked-dep` in binaries,
and pretty prints dependency tree in terminal with semver ranges
- `warg lock <namespace:package>`
Recursively fetches based on occurrences of `unlocked-dep` and creates a
"locked component". This locked component naively pins semver versions
and imports `locked-dep`s for those versions, and then instantiates each
component with the instantiation arguments that it expects. Given pinned
versions and integrity shas, this is intended to be easily reproducible.
- `warg bundle <namespace:package>`
An iteration on locking. For runtimes that are not registry aware (all
at this time), a locked component is not executable. This uses the warg
cache to replace `locked-dep` imports with the actual definitions of the
components being referenced, producing an executable artifact.

There are other import variants that are defined in the spec, that are
not currently being used, but are planned to be used in the future. For
more information on other available import use cases, see the
[explainer](https://github.com/WebAssembly/component-model/blob/main/design/mvp/Explainer.md#import-and-export-definitions)