Improved package init and publish flow (#269)
When publishing either a package `Release` or `Init`, a new error
`ConflictPendingPublish(RecordId)` may be returned.

This instructs the client that it could wait for the conflicting pending
record (on the same `LogId` package) to complete and then retry to
publish.

If a registry service offers key management and needs to sign the `Init`
record, this enables a smoother experience and messaging on the CLI.
Where the registry would return the `RecordId` of the registry signed
`Init` record as a `ConflictPendingPublish` error, the CLI would wait
for that record to publish and then retry. Since the package is already
initialized, an error message would be presented that indicates that the
package was initialized but not with the `Init` record that you signed.

Also, added a feature flag for `cli-interactive` for the `warg-client`
and a config flag to disable dialoguer interactive prompting even when
that feature is enabled.

Also, added paramenter to the `warg download` subcommand `--output` to write the downloaded `.wasm` file to a path in addition to the local cache.